// src/models/device.model.ts
export interface DeviceDetail {
  deviceId: number;
  deviceManagementListDataId: number;
  mac: string;
  deviceStatusId?: number;
  deviceStatusName: string;
  packingDate?: string;
  shipmentDate?: string;
  deviceRegisteredDate?: string;
  productId?: number;
  productName: string;
  siM1: string;
  sim1StatusId?: number;
  sim1Status: string;
  siM2: string;
  sim2StatusId?: number;
  sim2Status: string;
  customerAccountId: number;
  ncmGroupName: string;
  routerStatus: string;
  packageStatusId?: number;
  modemStatus: string;
  ethernetStatus: string;
  sim1ProvisioningDate?: string;
  sim2ProvisioningDate?: string;
  packageStatus: string;
  createdBy: string;
  createdOn?: string;
  lastUpdatedOn?: string;
  wifiStatus: string;
  eisStatus: string;
  deviceSerialNumber: string;
  wifiConfiguration: string;
  wifiConnectivityParam1: string;
  wifiConnectivityParam2: string;
  wifiConnectivityParam3: string;
  totalRows?: number;
}